refactor(all): 접근제어자 수정
- private으로 선언되어야하는데 테스트를 위해 public으로 선언한 경우
접근제어자를 protected로 수정하고 테스트클래스를 같은 패키지계층으로 이동
- private으로 선언 가능한 메소드 접근제어자 수정
