+++ src/test/java/racingcar/model/vo/EntryTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model.vo;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -7,8 +7,6 @@ import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.model.vo.Car;
-import racingcar.model.vo.Entry;
 
 public class EntryTest {
